Extract shared jsonc edit helpers
